refactor(cli): move the add group into its own module
Cuts `gpio add` out of cli/main.py into geoparquet_io/cli/commands/add.py,
unchanged. Clean cut: the group owns no helpers beyond its own commands.

Two test files resolved the `add_*_column_impl` aliases through
`geoparquet_io.cli.main` and now patch on the group module instead:
`tests/test_write_memory_forwarding.py` (six forwarding cases plus one
`patch.object`, which leaves it with no reference to `cli.main` at all) and
`tests/test_cli_api_call_parity_scaffold.py` (six `module=cli_add` cases).

Four pre-existing `arg-type` errors in `add kdtree` were suppressed by the
mypy per-module override on `geoparquet_io.cli.main`, so the override entry
follows the code rather than the errors being fixed here.

tests/data/cli_surface.json is byte-identical.
