+++ tests/test_cli_api_call_parity_scaffold.py
@@ -86,6 +86,7 @@ from click.testing import CliRunner
 from geoparquet_io.api import ops
 from geoparquet_io.api import table as table_module
 from geoparquet_io.cli import main as cli_main
+from geoparquet_io.cli.commands import add as cli_add
 from geoparquet_io.cli.commands import convert as cli_convert
 from geoparquet_io.cli.commands import sort as cli_sort
 from geoparquet_io.core import extract as core_extract
@@ -263,6 +264,7 @@ CASES: list[ParityCase] = [
             "add_bbox_column_impl",
             core_bbox.add_bbox_column,
             lambda c: ["add", "bbox", c.input_file, c.output_file],
+            module=cli_add,
         ),
         ops=_ops("add_bbox_table", core_bbox.add_bbox_table, lambda c: ops.add_bbox(c.table)),
         table=_table(
@@ -279,6 +281,7 @@ CASES: list[ParityCase] = [
             "add_h3_column_impl",
             core_h3.add_h3_column,
             lambda c: ["add", "h3", c.input_file, c.output_file],
+            module=cli_add,
         ),
         ops=_ops("add_h3_table", core_h3.add_h3_table, lambda c: ops.add_h3(c.table)),
         table=_table(
@@ -295,6 +298,7 @@ CASES: list[ParityCase] = [
             "add_s2_column_impl",
             core_s2.add_s2_column,
             lambda c: ["add", "s2", c.input_file, c.output_file],
+            module=cli_add,
         ),
         ops=_ops("add_s2_table", core_s2.add_s2_table, lambda c: ops.add_s2(c.table)),
         table=_table(
@@ -311,6 +315,7 @@ CASES: list[ParityCase] = [
             "add_a5_column_impl",
             core_a5.add_a5_column,
             lambda c: ["add", "a5", c.input_file, c.output_file],
+            module=cli_add,
         ),
         ops=_ops("add_a5_table", core_a5.add_a5_table, lambda c: ops.add_a5(c.table)),
         table=_table(
@@ -327,6 +332,7 @@ CASES: list[ParityCase] = [
             "add_quadkey_column_impl",
             core_quadkey.add_quadkey_column,
             lambda c: ["add", "quadkey", c.input_file, c.output_file],
+            module=cli_add,
         ),
         ops=_ops(
             "add_quadkey_table",
@@ -351,6 +357,7 @@ CASES: list[ParityCase] = [
             "add_kdtree_column_impl",
             core_kdtree.add_kdtree_column,
             lambda c: ["add", "kdtree", c.input_file, c.output_file],
+            module=cli_add,
         ),
         ops=_ops(
             "add_kdtree_table",
